Add toolchain file to specify nightly compiler
I didn't want to swap between nightly and stable constantly, and found
that there was a way to just set the compiler to nightly for a specific
project.

I know this isn't a super official repo, just a help repo, so not so
serious, but it makes it easier for any rando to use if need be :).